последний коммит

+++ CourseWorkHash/DivisionHashFunc.cs
@@ -5,19 +5,22 @@ using System.Text;
 
 namespace CourseWorkHash
 {
+    //Класс реализует хеш-функцию метода деления
     public class DivisionHashFunc : IHashFunc
     {
         public string Name => "Метод деления";
 
-        public int GetHash(string item, int size)
+        public long GetHash(string item, int size)
         {
-            int value = 0;
+            long value = 0;
 
+            //Полученная строка преобразуется в value сложением её символов
             for (int i = 0; i < item.Length; i++)
             {
                 value += item[i];
             }
 
+            //Берется остаток от деления получившегося значения на число ячеек в таблице 
             return value % size;
         }
     }
