+++ CourseWorkHash/MidSquareHashFunc.cs
@@ -5,14 +5,14 @@ using System.Text;
 
 namespace CourseWorkHash
 {
+    //Класс реализует хеш-функцию метода середины квадрата
     public class MidSquareHashFunc : IHashFunc
     {
         public string Name => "Метод середины квадрата";
 
-        public int GetHash(string item, int size)
+        public long GetHash(string item, int size)
         {
-            //Метод середины квадрата
-            int key = 0;
+            long key = 0;
             for (int i = 0; i < item.Length; i++)
             {
                 key += item[i];
@@ -22,6 +22,7 @@ namespace CourseWorkHash
             key *= key;
 
             int m = size;
+            //Сколько байт занимает количество ячеек size
             int nInBytes = 0;
             while (m != 0)
             {
@@ -29,6 +30,7 @@ namespace CourseWorkHash
                 m /= 2;
             }
 
+            //Вычисление сдвига, чтобы потом сделать побитовый сдвиг
             int shift;
             if (nInBytes >= 32)
             {
